requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
asyncpg==0.30.0
google-genai==1.24.0
pytz==2025.2
apscheduler==3.11.0
pandas==2.3.1
openpyxl==3.1.5
captcha==0.7.1